fsck before resize2fs (#253)
when we bumped the latest stable version of flatcar resizefs started
throwing an error to run fsck first. The tool suggests it. I have also
added the "-p" flag. Which will fix any errors it can without human
intervention. Non fixable errors will be cause the exit code to be
logically "or'd" with 4 to alert us of any problem.