refactor: revert functions in release tests
instead of undoing each step inside the test with defer functions
and named return values, return closures that will be called in
reverse order outside the test function. this makes sure that we
can pause on fail correctly even when there are steps to be taken
to undo the configuration changes, and hopefully makes the code
a little more readable.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>