feat(release-test): skip subinterfaces tests
add a new flag for tests that requires subinterfaces, and skip
them if any of the switches under test do not support them.
note that this is a pretty ham-fisted approach, we could instead
check whether the specific switch(es) that do not support them
are affected by each test, but the additional complexity of the
checks does not seem justified to me.

also move the check for an entirely skipped suite, so we do not
run the setup for nothing

Closes: #491

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>