chore: generate externals by default
generate 1 external by default. For external connections,
we will attempt to generate one using an ESLAG leaf if available,
else we will use an orphan leaf, and finally an MCLAG leaf
if no other option is available. This logic only kicks in if
externals are requested but no external connection is specified.

Note that we do not want MCLAG to be the default due to an
existing issue with external peering over those, and that
the collapsed core topology ONLY has MCLAG leaves. This is
the main reason why we are using this implicit logic, as
opposed to setting the eslag external connections default
value to 1.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>