fix(release): extra skip + fail count
the 'Multi-Subnets with filtering' is using local peering and
so can fail if subinterfaces are not supported, tag it
accordingly.

the total failure count at the end of the test was wrong if
some test had been skipped at execution time, due to us not
resetting the error (used for the skip message) and also
counting it as a failure.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>