fix: handle ext conns in hybrid envs
if a physical switch is connected to a hardware external, the
correponding ExternalConnection does not need to be created.
Since the connection does not specify the external, the addLink
function cannot know from the current config whether that
connection is physical or virtual. Work around this by also
annotating hardware ExternalConnections and ignoring them
when processing the virtual externals.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>