chore: always do curls in release tests
previously they were only enabled if the extended flag was set, to try
to reduce the amount of time it took. But without curls we are not
really testing external peerings, so we will have to live with it.
Also increase a little the time limit for ci jobs to compensate.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>