fix: support virt-externals with hw switches
the previous check was skipping external connections that were
hardware annotated, but that is incorrect. what we really want
is to perform the extra config steps (adding a link towards the
externals and the VRF information) only if the connection goes
to a virtual externals. the check is slightly complicated by the
fact that we need to iterate over the attachments, but now things
should work properly.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>