chore: ignore iperf speed for virtual switches
the minimal speed is a helpful tool to detect whether physical
switches are software-forwarding traffic, but for virtual
switches this does not apply. often release-tests on virtual
switches will fail with some iperf flow being slower than the
already low limit we were setting, and we have not done anything
in response to that anyway otehr than restarting the test, so stop
checking for that condition.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>