feat: test-connectivity flag all-servers
we are already skipping servers that are not attached to any
VPC as part of the checks to make test-connectivity work with
ESLAG and L3VNIs. However at the moment this prints a warning
re ESLAG, and it could simply be that the server has no
attachment. Regardless, there are situations when someone is
calling test-connectivity on a manual setup where only some
servers are attached to a VPC, and it is useful to be able to
do so. This PR introduces a flag (all-servers) which will make
the test fail if it detects that some server has no vpc attach,
and fixes the message when we skip a server due to this reason.
By default the flag is not set, and tests will simply skip
non-attached servers.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>