chore: make multiple ext. attach an error
the warning is easy to miss and then we end up spending a lot
of time troubleshooting this. I'm keeping the option to force
this with a --yes flag because we might still want to test this
in the future, or use the generated yaml as a base that will
be edited afterwards.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>