chore: smart default for gw uplinks
unless specified, vlab gen will now create as many gw
uplinks as spines (for spine-leaf topologies) or mesh
nodes (for mesh topologies), up to a maximum of 2, to
match the expected dual port card we are planning to
use. specifying a non-zero value from the command line
will override this default.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>